AdMapper,CommentMapper,UserMapper

+++ src/main/java/ru/skypro/homework/repository/AdRepository.java
@@ -3,6 +3,6 @@ package ru.skypro.homework.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.skypro.homework.model.Ad;
 
-public interface AdRepository extends CrudRepository<Ad, Long> {
+public interface AdRepository extends CrudRepository<Ad, Integer> {
 
 }
