+++ src/main/java/ru/skypro/homework/repository/CommentRepository.java
@@ -3,6 +3,6 @@ package ru.skypro.homework.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.skypro.homework.model.Comment;
 
-public interface CommentRepository extends CrudRepository<Comment, Long> {
+public interface CommentRepository extends CrudRepository<Comment, Integer> {
 
 }
